requirements.txt
torch>=2.0.0
torchvision
numpy
opencv-python
transformers
diffusers>=0.24.0
accelerate
peft
huggingface_hub
packaging
tqdm
omegaconf
datasets
lpips
Pillow
ftfy
scipy
kornia
torchsummary
